SS_for_logging_online_bank_account.java
Here's an example of an advanced security system implemented in Java for logging in to an online bank account.

In this example, the User class represents a user of the bank account. It has a username and a hashed password. The password is hashed using the SHA-256 algorithm. The authenticate method is used to compare the entered password (hashed) with the stored password hash.

The BankAccount class represents the online bank account. It has a users map to store the registered users. The registerUser method creates a new User object with the provided username and password, and stores it in the users map. The loginUser method retrieves the corresponding User object based on the username, and then calls the authenticate method to validate the entered password.

In the main method, an instance of the BankAccount class is created. The user is prompted to enter their username and password. The registerUser method is called to register a new user (you can modify this part to add more users). Finally, the loginUser method is called to attempt to log in the user with the entered credentials.

Note that this is a simplified example for demonstrating the security system. In a real-world scenario, you would typically have additional security measures, such as password complexity requirements, account locking after multiple failed login attempts, session management, and more.